requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
cached-property==1.3.0    # via regulations
certifi==2017.4.17        # via pyelasticsearch, requests
cfenv==0.5.3
chardet==3.0.4            # via requests
dj-database-url==0.4.2
django-haystack==2.6.1
django-mptt==0.8.7        # via regcore
django-overextends==0.4.3
django==1.11.3            # via django-haystack, django-overextends, regcore, regulations
elasticsearch==1.9.0      # via pyelasticsearch
enum34==1.1.6             # via regulations
furl==1.0.0               # via cfenv
futures==3.1.1            # via regulations
gunicorn==19.7.1
idna==2.5                 # via requests
jsonschema==2.6.0         # via regcore
newrelic==2.88.1.73
orderedmultidict==0.7.11  # via furl
psycopg2==2.7.3
pyelasticsearch==1.4
pytz==2017.2              # via django
pyyaml==3.12
regcore==3.1.0
regulations==8.0.0
requests==2.18.2          # via regulations
simplejson==3.10.0        # via pyelasticsearch
six==1.10.0               # via furl, orderedmultidict, pyelasticsearch, regcore, regulations
urllib3==1.22             # via elasticsearch, pyelasticsearch, requests
whitenoise==3.3.0
